Fix material property parameter values
